requirements.txt
streamlit
PyGithub
pandas
gspread 
google-auth 
google-auth-oauthlib 
google-auth-httplib2